requirements.txt
fastapi
uvicorn
requests
pydantic
huggingface-hub
python-dotenv
